ElasticSearchSpout processes results in separate thread; nextTuple and inQUery moved to abstract class
